+++ src/components/ActionMenu/components/Actions/Actions.tsx
@@ -1,15 +1,18 @@
-import React, {useCallback, useState} from 'react';
+import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import debounce from 'lodash/debounce';
 
 import {sortAndOverrideActionOrder} from '../../utilities';
 import {useFeatures} from '../../../../utilities/features';
+import {classNames} from '../../../../utilities/css';
 import type {
   MenuActionDescriptor,
   MenuGroupDescriptor,
 } from '../../../../types';
-import {Button} from '../../../Button';
 import {ButtonGroup} from '../../../ButtonGroup';
+import {EventListener} from '../../../EventListener';
 import {MenuGroup} from '../MenuGroup';
 import {MenuAction} from '../MenuAction';
+import {SecondaryAction} from '../SecondaryAction';
 
 import styles from './Actions.scss';
 
@@ -20,12 +23,30 @@ interface Props {
   groups?: MenuGroupDescriptor[];
 }
 
+const ACTION_SPACING = 4;
+
 export function Actions({actions = [], groups = []}: Props) {
   const {newDesignLanguage} = useFeatures();
+  const actionsLayoutRef = useRef<HTMLDivElement>(null);
+  const menuGroupRef = useRef<HTMLDivElement>(null);
+  const menuGroupWidthRef = useRef<number>(0);
+  const [actionWidths, setActionWidths] = useState<number[]>([]);
+  const [availableWidth, setAvailableWidth] = useState<number>(0);
   const [activeMenuGroup, setActiveMenuGroup] = useState<string | undefined>(
     undefined,
   );
+  const [showableActions, setShowableActions] = useState<
+    MenuActionDescriptor[] | null
+  >(null);
+  const [hiddenActions, setHiddenActions] = useState<MenuActionDescriptor[]>(
+    [],
+  );
 
+  const handleOffsetWidth = useCallback(
+    (width: number) =>
+      setActionWidths((actionWidths) => [...actionWidths, width]),
+    [],
+  );
   const handleMenuGroupToggle = useCallback(
     (group: string) => setActiveMenuGroup(activeMenuGroup ? undefined : group),
     [activeMenuGroup],
@@ -35,8 +56,61 @@ export function Actions({actions = [], groups = []}: Props) {
     () => setActiveMenuGroup(undefined),
     [],
   );
-  const menuActions = [...actions, ...groups];
 
+  const handleResize = useMemo(
+    () =>
+      debounce(
+        () => {
+          if (!newDesignLanguage || !actionsLayoutRef.current) return;
+          setAvailableWidth(actionsLayoutRef.current.offsetWidth);
+        },
+        20,
+        {leading: false, trailing: true, maxWait: 40},
+      ),
+    [newDesignLanguage],
+  );
+
+  useEffect(() => {
+    if (!actionsLayoutRef.current) return;
+    setAvailableWidth(actionsLayoutRef.current.offsetWidth);
+  }, [actionsLayoutRef]);
+
+  useEffect(() => {
+    if (!menuGroupRef.current) return;
+    menuGroupWidthRef.current = menuGroupRef.current?.offsetWidth;
+  }, [menuGroupRef]);
+
+  useEffect(() => {
+    if (!newDesignLanguage || actionWidths.length === 0 || availableWidth === 0)
+      return;
+
+    let currentAvailableWidth = availableWidth;
+    setShowableActions([]);
+    setHiddenActions([]);
+
+    actions.forEach((action, index) => {
+      if (
+        actionWidths[index] + menuGroupWidthRef.current + ACTION_SPACING <=
+        currentAvailableWidth
+      ) {
+        currentAvailableWidth -= actionWidths[index];
+        setShowableActions((showableActions) => [
+          ...(showableActions || []),
+          action,
+        ]);
+      } else {
+        currentAvailableWidth = 0;
+        setHiddenActions((hiddenActions) => [...hiddenActions, action]);
+      }
+    });
+  }, [actionWidths, actions, availableWidth, newDesignLanguage]);
+
+  const className = classNames(
+    styles.ActionsLayout,
+    newDesignLanguage && styles.newDesignLanguage,
+  );
+
+  const menuActions = [...actions, ...groups];
   const overriddenActions = sortAndOverrideActionOrder(menuActions);
 
   const actionMarkup = overriddenActions.map((action, index) => {
@@ -44,23 +118,29 @@ export function Actions({actions = [], groups = []}: Props) {
       const {title, actions, ...rest} = action;
 
       return actions.length > 0 ? (
-        <MenuGroup
-          key={`MenuGroup-${index}`}
-          title={title}
-          active={title === activeMenuGroup}
-          actions={actions}
-          {...rest}
-          onOpen={handleMenuGroupToggle}
-          onClose={handleMenuGroupClose}
-        />
+        <div key={`MenuGroup-${index}`} ref={menuGroupRef}>
+          <MenuGroup
+            title={title}
+            active={title === activeMenuGroup}
+            actions={[...hiddenActions, ...actions]}
+            {...rest}
+            onOpen={handleMenuGroupToggle}
+            onClose={handleMenuGroupClose}
+          />
+        </div>
       ) : null;
     }
-
+    if (showableActions && showableActions.length >= 0) return null;
     const {content, onAction, ...rest} = action;
     return newDesignLanguage ? (
-      <Button key={index} onClick={onAction} {...rest}>
+      <SecondaryAction
+        key={index}
+        onClick={onAction}
+        {...rest}
+        getOffsetWidth={handleOffsetWidth}
+      >
         {content}
-      </Button>
+      </SecondaryAction>
     ) : (
       <MenuAction
         key={`MenuAction-${index}`}
@@ -72,12 +152,26 @@ export function Actions({actions = [], groups = []}: Props) {
   });
 
   return (
-    <div className={styles.ActionsLayout}>
+    <div className={className} ref={actionsLayoutRef}>
       {newDesignLanguage ? (
-        <ButtonGroup>{actionMarkup}</ButtonGroup>
+        <ButtonGroup spacing="extraTight">
+          {showableActions && showableActions.length > 0
+            ? showableActions.map((action) => (
+                <SecondaryAction
+                  key={action.content}
+                  {...action}
+                  getOffsetWidth={handleOffsetWidth}
+                >
+                  {action.content}
+                </SecondaryAction>
+              ))
+            : null}
+          {actionMarkup}
+        </ButtonGroup>
       ) : (
         actionMarkup
       )}
+      <EventListener event="resize" handler={handleResize} />
     </div>
   );
 }