+++ src/components/ActionMenu/components/RollupActions/RollupActions.tsx
@@ -8,6 +8,7 @@ import type {
 import {useI18n} from '../../../../utilities/i18n';
 import {useToggle} from '../../../../utilities/use-toggle';
 import {useFeatures} from '../../../../utilities/features';
+import {classNames} from '../../../../utilities/css';
 import {ActionList} from '../../../ActionList';
 import {Button} from '../../../Button';
 import {Popover} from '../../../Popover';
@@ -24,6 +25,10 @@ export interface RollupActionsProps {
 export function RollupActions({items = [], sections = []}: RollupActionsProps) {
   const i18n = useI18n();
   const {newDesignLanguage} = useFeatures();
+  const classname = classNames(
+    styles.RollupActivator,
+    newDesignLanguage && styles.newDesignLanguage,
+  );
 
   const {value: rollupOpen, toggle: toggleRollupOpen} = useToggle(false);
 
@@ -32,9 +37,10 @@ export function RollupActions({items = [], sections = []}: RollupActionsProps) {
   }
 
   const activatorMarkup = (
-    <div className={styles.RollupActivator}>
+    <div className={classname}>
       <Button
         plain={!newDesignLanguage}
+        outline={newDesignLanguage}
         icon={HorizontalDotsMinor}
         accessibilityLabel={i18n.translate(
           'Polaris.ActionMenu.RollupActions.rollupButton',