+++ src/components/ButtonGroup/ButtonGroup.tsx
@@ -6,7 +6,11 @@ import {elementChildren} from '../../utilities/components';
 import {Item} from './components';
 import styles from './ButtonGroup.scss';
 
+type Spacing = 'extraTight' | 'tight' | 'loose';
+
 export interface ButtonGroupProps {
+  /** Determines the space between button group items */
+  spacing?: Spacing;
   /** Join buttons as segmented group */
   segmented?: boolean;
   /** Buttons will stretch/shrink to occupy the full width */
@@ -19,12 +23,14 @@ export interface ButtonGroupProps {
 
 export function ButtonGroup({
   children,
+  spacing,
   segmented,
   fullWidth,
   connectedTop,
 }: ButtonGroupProps) {
   const className = classNames(
     styles.ButtonGroup,
+    spacing && styles[spacing],
     segmented && styles.segmented,
     fullWidth && styles.fullWidth,
   );