+++ src/components/Page/components/Header/Header.tsx
@@ -2,6 +2,7 @@ import React, {isValidElement} from 'react';
 
 import {classNames} from '../../../../utilities/css';
 import {buttonsFrom} from '../../../Button';
+import {TextStyle} from '../../../TextStyle';
 import {useMediaQuery} from '../../../../utilities/media-query';
 import {useFeatures} from '../../../../utilities/features';
 import {
@@ -52,6 +53,8 @@ export interface HeaderProps extends TitleProps {
   actionGroups?: MenuGroupDescriptor[];
   /** Additional navigation markup */
   additionalNavigation?: React.ReactNode;
+  // Additional meta data
+  additionalMetaData?: React.ReactNode | string;
 }
 
 export function isPrimaryAction(
@@ -64,6 +67,7 @@ export function Header({
   title,
   subtitle,
   titleMetadata,
+  additionalMetaData,
   thumbnail,
   titleHidden = false,
   separator,
@@ -112,12 +116,14 @@ export function Header({
     ) : null;
 
   const pageTitleMarkup = (
-    <Title
-      title={title}
-      subtitle={subtitle}
-      titleMetadata={titleMetadata}
-      thumbnail={thumbnail}
-    />
+    <div className={styles.TitleWrapper}>
+      <Title
+        title={title}
+        subtitle={subtitle}
+        titleMetadata={titleMetadata}
+        thumbnail={thumbnail}
+      />
+    </div>
   );
 
   const primaryActionMarkup = primaryAction ? (
@@ -140,6 +146,12 @@ export function Header({
       </ConditionalWrapper>
     ) : null;
 
+  const additionalMetaDataMarkup = additionalMetaData ? (
+    <div className={styles.AdditionalMetaData}>
+      <TextStyle variation="subdued">{additionalMetaData}</TextStyle>
+    </div>
+  ) : null;
+
   const headerClassNames = classNames(
     styles.Header,
     titleHidden && styles.titleHidden,
@@ -154,6 +166,7 @@ export function Header({
     const {slot1, slot2, slot3, slot4, slot5, slot6} = determineLayout({
       breadcrumbMarkup,
       pageTitleMarkup,
+      additionalMetaDataMarkup,
       paginationMarkup,
       actionMenuMarkup,
       primaryActionMarkup,
@@ -175,7 +188,13 @@ export function Header({
               <div className={styles.RightAlign}>
                 <ConditionalWrapper
                   condition={[slot3, slot4].every(notNull)}
-                  wrapper={(children) => <ButtonGroup>{children}</ButtonGroup>}
+                  wrapper={(children) =>
+                    newDesignLanguage ? (
+                      <div className={styles.Actions}>{children}</div>
+                    ) : (
+                      <ButtonGroup>{children}</ButtonGroup>
+                    )
+                  }
                 >
                   {slot3}
                   {slot4}
@@ -236,16 +255,7 @@ function PrimaryActionMarkup({
     );
   }
 
-  return (
-    <ConditionalWrapper
-      condition={newDesignLanguage === false}
-      wrapper={(children) => (
-        <div className={styles.PrimaryActionWrapper}>{children}</div>
-      )}
-    >
-      {content}
-    </ConditionalWrapper>
-  );
+  return <div className={styles.PrimaryActionWrapper}>{content}</div>;
 }
 
 function shouldShowIconOnly(
@@ -279,6 +289,7 @@ function determineLayout({
   breadcrumbMarkup,
   pageTitleMarkup,
   title,
+  additionalMetaDataMarkup,
   paginationMarkup,
   actionMenuMarkup,
   primaryActionMarkup,
@@ -287,6 +298,7 @@ function determineLayout({
   breadcrumbMarkup: MaybeJSX;
   pageTitleMarkup: JSX.Element;
   title?: string;
+  additionalMetaDataMarkup: MaybeJSX;
   paginationMarkup: MaybeJSX;
   actionMenuMarkup: MaybeJSX;
   primaryActionMarkup: MaybeJSX;
@@ -309,7 +321,7 @@ function determineLayout({
         slot2: pageTitleMarkup,
         slot3: actionMenuMarkup,
         slot4: primaryActionMarkup,
-        slot5: null,
+        slot5: additionalMetaDataMarkup,
         slot6: null,
       },
       condition:
@@ -321,10 +333,10 @@ function determineLayout({
     mobileDefault: {
       slots: {
         slot1: breadcrumbMarkup,
-        slot2: null,
+        slot2: pageTitleMarkup,
         slot3: actionMenuMarkup,
         slot4: primaryActionMarkup,
-        slot5: pageTitleMarkup,
+        slot5: additionalMetaDataMarkup,
         slot6: null,
       },
       condition: isNavigationCollapsed,
@@ -335,7 +347,7 @@ function determineLayout({
         slot2: pageTitleMarkup,
         slot3: null,
         slot4: primaryActionMarkup,
-        slot5: null,
+        slot5: additionalMetaDataMarkup,
         slot6: null,
       },
       condition:
@@ -349,10 +361,15 @@ function determineLayout({
       slots: {
         slot1: breadcrumbMarkup,
         slot2: pageTitleMarkup,
-        slot3: null,
+        slot3: (
+          <>
+            {actionMenuMarkup}
+            {primaryActionMarkup}
+          </>
+        ),
         slot4: paginationMarkup,
-        slot5: actionMenuMarkup,
-        slot6: primaryActionMarkup,
+        slot5: additionalMetaDataMarkup,
+        slot6: null,
       },
       condition: !isNavigationCollapsed,
     },