+++ src/components/Page/components/Header/components/Title/Title.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 
-import {classNames} from '../../../../../../utilities/css';
+import {DisplayText} from '../../../../../DisplayText';
+import {TextStyle} from '../../../../../TextStyle';
 import type {AvatarProps} from '../../../../../Avatar';
 import type {ThumbnailProps} from '../../../../../Thumbnail';
-import {DisplayText} from '../../../../../DisplayText';
+import {classNames} from '../../../../../../utilities/css';
 import {useFeatures} from '../../../../../../utilities/features';
+import {useMediaQuery} from '../../../../../../utilities/media-query';
 
 import styles from './Title.scss';
 
@@ -23,10 +25,15 @@ export interface TitleProps {
 
 export function Title({title, subtitle, titleMetadata, thumbnail}: TitleProps) {
   const {newDesignLanguage} = useFeatures();
+  const {isNavigationCollapsed} = useMediaQuery();
+  const displaySize = isNavigationCollapsed ? 'large' : 'small';
   const titleMarkup = title ? (
     <div className={styles.Title}>
-      <DisplayText size="large" element="h1">
-        {title}
+      <DisplayText
+        size={newDesignLanguage ? displaySize : 'large'}
+        element="h1"
+      >
+        <TextStyle variation="strong">{title}</TextStyle>
       </DisplayText>
     </div>
   ) : null;